Adicionar aba com todas as atividades no admin dashboard
